upd: docs & tidying
