+++ examples/res/adder.json
@@ -1,5 +1,5 @@
 {
-  "schema_version": "0.3.0",
+  "schema_version": "0.4.0",
   "metadata": {
     "name": "abi",
     "version": "0.1.0",
